FROMLIST: mfd: cros_ec_i2c: moving the system sleep pm ops to late
The cros_ec_i2c driver is still active after it had suspended or before it
resumes. Besides that, it also tried to transfer data even after the I2C
host had been suspended. This will lead the system to crash.

During the test, we also observe that the EC needs to be resumed earlier
due to some status polling from the EC FW (e.g. battery status). So we
move the PM ops to late stage to make it work normally.

BUG=chromium:785506
TEST=Compile

Link: https://patchwork.kernel.org/patch/10241989/mbox/

Change-Id: Iede849324e542e319bac9300ab5a563fd979d7ca
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: Andy Shevchenko <andy.shevchenko@gmail.com>