BACKPORT: irqchip/pdc: Add PDC interrupt controller for QCOM SoCs
The Power Domain Controller (PDC) on QTI SoCs like SDM845 houses an
interrupt controller along with other domain control functions to handle
interrupt related functions like handle falling edge or active low which
are not detected at the GIC and handle wakeup interrupts.

The interrupt controller is on an always-on domain for the purpose of
waking up the processor. Only a subset of the processor's interrupts are
routed through the PDC to the GIC. The PDC powers on the processors'
domain, when in low power mode and replays pending interrupts so the GIC
may wake up the processor.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>

Conflicts arose because of additional unrelated drivers added
upstream.

Conflicts:
	drivers/irqchip/Kconfig
	drivers/irqchip/Makefile

BUG=b:110423885
TEST=Build kernel for Cheza

Signed-off-by: Evan Green <evgreen@chromium.org>
(cherry picked from commit f55c73aef8904ca41ff99618a3a9cf538eb88626)
Change-Id: I5001afc5dfd37ccf95fb3f8daf4b8a43973bc517
Reviewed-on: https://chromium-review.googlesource.com/1110673
Reviewed-by: Douglas Anderson <dianders@chromium.org>