UPSTREAM: drm/i915/gen9, gen10: Disable FBC on planes with a misaligned Y-offset
Enabling FBC on a plane having a Y-offset that isn't divisible by 4 may
cause pipe FIFO underruns and flickers, so disable FBC on such a config.

I tried the followings to work around the issue:
- enable each HW work around in ILK_DPFC_CHICKEN
- disable each compression algorithm in ILK_DPFC_CONTROL
- disable low-power watermarks
None of the above got rid of the problem. I haven't found this issue in
the Bspec/WA database either.

Besides the igt testcase below (yet to be merged) an easy way to
reproduce the issue is to enable a plane with FBC and a plane Y-offset
not aligned to 4 and then just enable/disable FBC in a loop, keeping the
plane enabled.

I could trigger the problem on BXT/GLK/SKL/CNL, so assume for now that it's
only present on GEN9 and GEN10.

v2: (Ville)
- Run the test/apply the WA on CNL as well.
- Use IS_GEN() instead of INTEL_GEN().
- Fix spelling.

BUG= None
TEST= Run IGT tests with below commnads, test result should be success
./kms_frontbuffer_tracking --run-subtest fbc-indfb-scaledprimary
./kms_frontbuffer_tracking --run-subtest fbc-shrfb-scaledprimary

Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Cc: Ville Syrjl <ville.syrjala@linux.intel.com>
Testcase: igt/kms_plane/plane-clipping-pipe-A-planes
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180301134457.13974-1-imre.deak@intel.com
(cherry picked from commit fee0fddc1dc114039bf9aed9fdeeecdc1949cdc7)
Signed-off-by: Kiran Kumar S <kiran.s.kumar@intel.corp-partner.google.com>

Change-Id: I9921f781f7c6dec472ae5f1a4430e13745640da2
Reviewed-on: https://chromium-review.googlesource.com/1388409
Commit-Ready: Kiran Kumar <kiran.s.kumar@intel.corp-partner.google.com>
Tested-by: Anand Naik <anand.naik@intel.corp-partner.google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>