UPSTREAM: ASoC: rt5651: Allow specifying over-current threshold through a device-property
OVer-Current-Detection (OVCD) for the micbias current is used to detect
if an inserted jack is a headset or headphones (mic shorted to ground).

Some boards may need different values for the OVCD current threshold
because of a resistor on the board in serial with or parallel to the
jack mic contact.

This commit adds support for configuring the OCVD current threshold
through the "realtek,over-current-threshold-microamp" device-property.

Note this commit changes the default value from 600uA to 2000uA,
because testing has shown 600uA to be a poor default.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 583a9debd71e21a317b1fc7b293c22c49b33d9e4)
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>

BUG=b:110025789
TEST=Test Audio use cases with full SOF 0.6 rebase release series applied.

Change-Id: I806436d563a3bc2524387912ab747f702e7f93d4
Reviewed-on: https://chromium-review.googlesource.com/1353455
Commit-Ready: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>