UPSTREAM: ASoC: Intel: bytcr_rt5651: Add quirk micbias OVCD configuration
Add support for setting the micbias OVCD limits device-properties through
quirks.

And set the limits for this to 2000uA with a scale-factor of 0.75 for the
KIANO SlimNote 14.2 device, which is the only device on which
jack-detection is currently enabled.

Tested-by: Carlo Caione <carlo@endlessm.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 8ffaa6a136e6f30e053e78fb9742c7748bef8576)
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>

BUG=b:110025789
TEST=Test Audio use cases with full SOF 0.6 rebase release series applied.

Change-Id: I29d8e1920048e21551ec44bc0d18a941f52967ff
Reviewed-on: https://chromium-review.googlesource.com/1353465
Commit-Ready: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Kamal Ap <kamal.ap@intel.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>