FROMGIT: arm64/debug: Fix registers on sleeping tasks
This is the equivalent of commit 001bf455d206 ("ARM: 8428/1: kgdb: Fix
registers on sleeping tasks") but for arm64.  Nuff said.

...well, perhaps I could also add that task_pt_regs are userspace
registers and that's not what kgdb is supposed to be reporting.  We're
supposed to be reporting kernel registers.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>

BUG=chromium:821174
TEST='info thread' in kgdb now works for sleeping tasks

Change-Id: I0e995418f4947c7ce6f6e803126fbe711d3aa0a6
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 24153c03d4f5dc4ac0ccd446a9636c269c1d4a02
 git://git.kernel.org/pub/scm/linux/kernel/git/arm64/linux.git for-next/core)
Reviewed-on: https://chromium-review.googlesource.com/959393
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>