FIXUP: FROMLIST: clk: qcom: Clear hardware clock control bit of RCG
Note: This is a trivial fixup patch to upgrade from v1 to v2 of the
series.

For upcoming targets like sdm845, POR value of the hardware clock control
bit is set for most of root clocks which needs to be cleared for software
to be able to control. For older targets like MSM8996, this bit is reserved
bit and having POR value as 0 so this patch will work for the older targets
too. So update the configuration mask to take care of the same to clear
hardware clock control bit.

Signed-off-by: Amit Nischal <anischal@codeaurora.org>

BUG=chromium:795946
TEST=Build kernel for Cheza

Signed-off-by: Evan Green <evgreen@chromium.org>
(am from https://patchwork.kernel.org/patch/10267105/)
Change-Id: I016bea9fc4fb3f579495247b78d9a35e5a0160e8
Reviewed-on: https://chromium-review.googlesource.com/956101
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>