UPSTREAM: ASoC: da7219: Add common clock usage for providing DAI clks
Reland

commit 6b27cd601234e ("BACKPORT: FROMGIT: ASoC: da7219: Add common clock usage for providing DAI clks")

The difference is that now we are using component in codec driver.

There is a need to use DA7219 as DAI clock master for other codecs
within a system, which means that the DAI clocks are required to
remain, regardless of whether the codec is actually running
playback/capture. To be able to expose control of the DAI clocking
the common clock framework has been employed.

The current implementation adds a simple clock gate for enabling
and disabling the DAI clocks, with no rate control supported
(this is still handled through standard hw_params() functions as
before). If DT is enabled then the clock is added to the OF
providers list, otherwise a clkdev lookup is used.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit fc8f7ea2d6c074baaad202c9187962bfa493ef13)

BUG=b:80592327, b:73825035
TEST=ARCH=arm64 make allyesconfig;
     CROSS_COMPILE=aarch64-cros-linux-gnu- ARCH=arm64 make -j64

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I2f58c04569fffc763033fceb0bee6b18a74c45ba
Reviewed-on: https://chromium-review.googlesource.com/1139724