UPSTREAM: ASoC: da7219: clkdev_drop usage depends on CONFIG_COMMON_CLK
Fixes: ASoC: da7219: Add common clock usage for providing DAI clks

clkdev_drop usage in the codec remove function should be dependent
on if CONFIG_COMMON_CLK is defined for the platform, otherwise it
can cause build failures for platforms that do not support this.
The clkdev_* functions are still defined for those platforms, in
headers and source but the functions are not linked in.

This patch resolves this issue, so clkdev_drop is only used if
CONFIG_COMMON_CLK is defined.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 0f9d0e355d7c3fa5756313765a77961074017845)

BUG=b:80592327, b:73825035
TEST=ARCH=arm64 make allyesconfig;
     CROSS_COMPILE=aarch64-cros-linux-gnu- ARCH=arm64 make -j64

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Ice798f85f2a42cc4f527ce1a2ea8741080c36967
Reviewed-on: https://chromium-review.googlesource.com/1139725