CHROMIUM: config: enable CONFIG_SCSI_SCAN_ASYNC
In CL:965870 we decided that we wanted to enable
CONFIG_SCSI_SCAN_ASYNC for 64-bit Rockchip devices.  There should be
nothing Rockchip-specific about that config option, but it seems
unwise to enable it in across all architectures without giving it
enough bake time.  Thus the config was not made to base.config in 4.4.

In kernel 4.14 it seems like we should enable this option in general.
Presumably if it causes any issues we still have time to debug them.

This is the equivalent of
    echo 'CONFIG_SCSI_SCAN_ASYNC=y' >> chromeos/config/base.config
    ./chromeos/scripts/kernelconfig olddefconfig
...atop an already normalized config.

BUG=None
TEST=Build and boot

Change-Id: I52570557cb042d98f0305d22b840e5aeb2d94053
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/966596
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>