BACKPORT: gpiolib: Support 'gpio-reserved-ranges' property
Some qcom platforms make some GPIOs or pins unavailable for use by
non-secure operating systems, and thus reading or writing the registers
for those pins will cause access control issues. Add support for a DT
property to describe the set of GPIOs that are available for use so that
higher level OSes are able to know what pins to avoid reading/writing.
Non-DT platforms can add support by directly updating the
chip->valid_mask.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Timur Tabi <timur@codeaurora.org>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
(cherry picked from commit 726cb3ba49692bdae6caff457755e7cdb432efa4)

Conflicts:
   drivers/gpio/gpiolib.c
...upstream gpiochip_add_irqchip() has more parameters and some of the
code added in this patch was adjacent to a call to
gpiochip_add_irqchip().  Thus the merge conflict.  Picking the whole
series of patches to add more paremters to gpiochip_add_irqchip() was
a bit too invasive and since this was just a conflict for a nearby
line of code, not picking a tons of extra patches to avoid the
conflict.

BUG=b:74036910
TEST=Set reserved-gpios in DT, saw they couldn't be requested

Change-Id: I0864ef623a8b77c8742037039fba27d959a7f1ef
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1105292
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>