CHROMIUM: iwl7000: mac80211: Fix setting of SAE authentication timeout
In commit 69cc9c8821a9 ("mac80211: Adjust SAE authentication timeout")
the SAE timeout was adjusted to adhere with the new specification
definitions. However, this did not consider the case that the driver
does not support Tx status report, and did not consider the fact that
in case that Tx status report is indeed supported, the timeout is reset
to a lower value.

Fix these issues by only setting the SAE timeout for the case that Tx
status is not supported, and in addition, in case that the Tx status
is supported, set the SAE timeout only after the frame is ACKed.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 3dd603c9e9e1b3dee46b42d54a22bddf34c83df8