CHROMIUM: iwl7000: mac80211: remove check for local->wolan in gtk_rekey_add()
We call ieee80211_gtk_rekey_add() when "resuming" from the d3_test in
iwlmvm, if there was a rekeying while pseudo-suspended.  The
ieee80211_gtk_rekey_add() in mac80211 warns if local->wowlan is not
set (which is the case with d3_test), so waking up after rekeying
doesn't work.

We could officially remove the check in mac80211, but it could cause
problems if gtk_rekey_notify() is not called immediately after
gtk_rekey_add() was called.  So the right solution would be to
refactor these two functions so that they are together to prevent such
issues.  For now (since I don't have much spare time :P) just hack the
check out and mark it as NOUPSTREAM.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: ee76106ce442cbae1063ed7620e8a719a87b50a2