FROMGIT: pwm-backlight: Enable/disable the PWM before/after LCD enable toggle.
Before this patch the enable signal was set before the PWM signal and
vice-versa on power off. This sequence is wrong, at least, it is on
the different panels datasheets that I checked, so I inverted the sequence
to follow the specs.

For reference the following panels have the mentioned sequence:
  - N133HSE-EA1 (Innolux)
  - N116BGE (Innolux)
  - N156BGE-L21 (Innolux)
  - B101EAN0 (Auo)
  - B101AW03 (Auo)
  - LTN101NT05 (Samsung)
  - CLAA101WA01A (Chunghwa)

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Acked-by: Daniel Thompson <daniel.thompson@linaro.org>
Acked-by: Jingoo Han <jingoohan1@gmail.com>
Acked-by: Thierry Reding <thierry.reding@gmail.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>

BUG=b:80590153
TEST=Build kernel for Cheza

Signed-off-by: Evan Green <evgreen@chromium.org>
(cherry picked from commit 5fb5caee92ba35a4a3baa61d45a78eb057e2c031
 https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git master)
Change-Id: If69d4376fac0ed0d6dd35786117c7bf8e7c0da14
Reviewed-on: https://chromium-review.googlesource.com/1112393
Reviewed-by: Douglas Anderson <dianders@chromium.org>