UPSTREAM: z3fold: limit use of stale list for allocation
Currently if z3fold couldn't find an unbuddied page it would first try
to pull a page off the stale list.  The problem with this approach is
that we can't 100% guarantee that the page is not processed by the
workqueue thread at the same time unless we run cancel_work_sync() on
it, which we can't do if we're in an atomic context.  So let's just
limit stale list usage to non-atomic contexts only.

Link: http://lkml.kernel.org/r/47ab51e7-e9c1-d30e-ab17-f734dbc3abce@gmail.com
Signed-off-by: Vitaly Vul <vitaly.vul@sony.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: <Oleksiy.Avramchenko@sony.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

BUG=chromium:822360
TEST=Build and boot

Change-Id: Ib2260339af06cf88d19c26751fd32de3fc75cb7e
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 5c9bab592f53328b3e00d7293378b714abf47a2a)
Reviewed-on: https://chromium-review.googlesource.com/1017011
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>