FROMGIT: clk: qcom: Base rcg parent rate off plan frequency
_freq_tbl_determine_rate uses the pre_div found in the clock plan
multiplied by the requested rate from the caller to determine the
best parent rate to set. If the requested rate is not exactly equal
to the rate that was found in the clock plan, then using the requested
rate in parent rate calculations is incorrect. For instance, if 150MHz
was requested, but 200MHz was the match found, and that plan had a
pre_div of 3, then the parent should be set to 600MHz, not 450MHz.

Signed-off-by: Evan Green <evgreen@chromium.org>
Fixes: bcd61c0f535a ("clk: qcom: Add support for root clock generators (RCGs)")
Signed-off-by: Stephen Boyd <sboyd@kernel.org>

BUG=b:78656924
TEST=Build kernel for Cheza

Change-Id: I4959becd24d76b4ce6b12e58d2fc7c54aa8a9eb9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit c7d2a0eb6c028ba064bfe92d7667977418142c7c
 git://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-next)
Reviewed-on: https://chromium-review.googlesource.com/1053368