FROMGIT: regulator: Don't return or expect -errno from of_map_mode()
In of_get_regulation_constraints() we were taking the result of
of_map_mode() (an unsigned int) and assigning it to an int.  We were
then checking whether this value was -EINVAL.  Some implementers of
of_map_mode() were returning -EINVAL (even though the return type of
their function needed to be unsigned int) because they needed to
signal an error back to of_get_regulation_constraints().

In general in the regulator framework the mode is always referred to
as an unsigned int.  While we could fix this to be a signed int (the
highest value we store in there right now is 0x8), it's actually
pretty clean to just define the regulator mode 0x0 (the lack of any
bits set) as an invalid mode.  Let's do that.

Fixes: 5e5e3a42c653 ("regulator: of: Add support for parsing initial and suspend modes")
Suggested-by: Javier Martinez Canillas <javierm@redhat.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

BUG=b:74988960
TEST=Build and boot

Change-Id: I580a16fc098cf92c476cd1d19c5559a17b1e03d3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 02f3703934a42417021405ef336fe45add13c3d1
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/regulator.git for-4.18)
Reviewed-on: https://chromium-review.googlesource.com/1022505
Reviewed-by: Alexandru M Stan <amstan@chromium.org>