FROMGIT: soc: qcom: cmd-db: Make endian-agnostic
This driver deals with memory that is stored in little-endian format.
Update the structures with the proper little-endian types and then
do the proper conversions when reading the fields. Note that we compare
the ids with a memcmp() because we already pad out the string 'id' field
to exactly 8 bytes with the strncpy() onto the stack.

Cc: Mahesh Sivasubramanian <msivasub@codeaurora.org>
Cc: Lina Iyer <ilina@codeaurora.org>
Cc: Bjorn Andersson <bjorn.andersson@linaro.org>
Cc: Evan Green <evgreen@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>

BUG=b:74981586
TEST=Boot kernel for Cheza

Change-Id: Ia08fec756df20be6e4b7bea022afef81bcaa1760
Signed-off-by: Evan Green <evgreen@chromium.org>
(cherry picked from commit 163aa82692e70eb631d6627ade286f9b89bdc162
 git://git.kernel.org/pub/scm/linux/kernel/git/agross/linux.git for-next)
Reviewed-on: https://chromium-review.googlesource.com/1005447
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>