FROMLIST: ACPI: APD: Add AMD misc clock handler support
AMD SoC exposes clock for general purpose use. The clock registration
is done in clk-st driver. The MMIO mapping are passed on to the
clock driver for accessing the registers.
The misc clock handler will create MMIO mappings to access the
clock registers and enable the clock driver to expose the clock
for use of drivers which will connect to it.

Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

(am from https://patchwork.kernel.org/patch/10388975/)

BUG=b:74570989
TEST=aplay -vv <file>
  check register to see clock enabled
  kill aplay
  check register to see clock disabled

Change-Id: Ie09eccb461ce051ddbdf0a0c131abcadca352b35
Reviewed-on: https://chromium-review.googlesource.com/1023513
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>