FROMGIT: clk: qcom: gdsc: Add support to poll CFG register to check GDSC state
The default behavior of the GDSC enable/disable sequence is to
poll the status bits of either the actual GDSCR or the
corresponding HW_CTRL registers.

On targets which have support for a CFG_GDSCR register, the
status bits might not show the correct state of the GDSC,
especially in the disable sequence, where the status bit
will be cleared even before the core is completely power
collapsed. On targets with this issue, poll the power on/off
bits in the CFG_GDSCR register instead to correctly determine
the GDSC state.

Signed-off-by: Amit Nischal <anischal@codeaurora.org>
Signed-off-by: Taniya Das <tdas@codeaurora.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>

BUG=b:78656924
TEST=With more patches atop, boot sdm845-mtp to UI

Change-Id: I3363bb5ecb88f863b2dee19227b37615a4b1677a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit e892e17d0c0e3d1bb8a56b84d15f9995807b92fa
 git://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-next)
Reviewed-on: https://chromium-review.googlesource.com/1031104
Reviewed-by: Evan Green <evgreen@chromium.org>