FROMGIT: clk: qcom: Simplify gdsc status checking logic
The code is complicated because we want to check if the GDSC is enabled
or disabled based on different bits in different registers while the
GDSC hardware is slightly different across chips. Furthermore, we poll
the status of the enable or disable state by checking if the gdsc is
enabled or not, and then comparing that to if the gdsc is being enabled
or disabled. Let's push all that into one function, so we can ask if the
status matches what we want, either on or off. Then the call site can
just ask that question, and the logic to check that state can simply
return yes or no, and not 1 or 0 or 0 or 1 depending on if we're
enabling or disabling respectively.

Tested-by: Taniya Das <tdas@codeaurora.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>

BUG=b:78656924
TEST=With more patches atop, boot sdm845-mtp to UI

Change-Id: I9e9c33af31521b21fcc2da1ab1241fb7ecd96e82
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 88051f55de502811472fe25d4a710bc7e6e60fd1
 git://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-next)
Reviewed-on: https://chromium-review.googlesource.com/1053118
Reviewed-by: Evan Green <evgreen@chromium.org>