FROMLIST: arm64: dts: qcom: sdm845: Sort nodes in the soc by address
This is pure-churn and should be a no-op.  I'm doing it in the hopes
of reducing merge conflicts.  When things are sorted in a sane way
(and by base address seems sane) then it's less likely that future
patches will cause merge conflicts.

Signed-off-by: Douglas Anderson <dianders@chromium.org>

BUG=chromium:795946
TEST=Everything works the same

Change-Id: If6ae31eafb9c11b20e7c0026263e04629ed78b87
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/10381745/)
Reviewed-on: https://chromium-review.googlesource.com/1054775
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>