CHROMIUM: drm/i915: Wait for vblank after disabling cursor
The cursor disable write is latched on vsync but if we do a modeset
before that happens the write seems to be discarded and the cursor
stays on. The CURCNTL register will still read back 0 as if the write
took, but underlying hardware didn't change. We don't disable the
cursor often, so we can just force the pre-GEN9 path for cursor
that waits for vblank.

BUG=b:70546200, b:70534167
TEST=Toggle primary/extended display as described in bug

(cherry picked from commit 24437c01ae51d2f3eaa301a8c901e039a5646c86)
https://chromium-review.googlesource.com/#/c/chromiumos/third_party/kernel/+/877396/

Signed-off-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Sean Paul <seanpaul@google.com>

Change-Id: Ie53c3abef510b2fcf34a12b440bd906c70c03843
Signed-off-by: Gaurav K Singh <gaurav.k.singh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1055203
Reviewed-by: Justin TerAvest <teravest@chromium.org>