UPSTREAM: ASoC: Intel: bytcr_rt5640: Set card long_name based on quirks
Many X86 devices using a BYT SoC + RT5640 codec are cheap devices with
generic DMI strings, causing snd_soc_set_dmi_name() to fail to set a
long_name, making it impossible for userspace to have a correct UCM
profile which only uses inputs / outputs which are actually hooked up
on the device.

Our quirks already specify which input the internal mic is connected to
and if a single (mono) speaker is used or if the device has stereo
speakers.

This commit sets a long_name based on the quirks so that userspace can
have UCM profiles doing the right thing based on the long_name.

Note that if we ever encounter the need for a special UCM profile for
some device we can add a quirk to set a specific long_name for the
device,

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Acked-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 063422ca2a9de238401c3848c1b3641c07b6316c)
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>

BUG=b:110025789
TEST=Test Audio use cases with full SOF 0.6 rebase release series applied.

Change-Id: Id4a563d7d3db4b730fc08e07c4728338e9c316d6
Reviewed-on: https://chromium-review.googlesource.com/1353567
Commit-Ready: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Kamal Ap <kamal.ap@intel.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>