CHROMIUM: Revert "FROMGIT: arm64: dts: qcom: sdm845: Fixup clocks to use raw values"
This change reverts and fixes a previous change that used raw
clock numbers in order to land the device tree before the clock
driver had landed. Since that change was made, the clock numbers
have changed, so in addition to being good housekeeping, it's
needed in order to boot.

Apparently, the upstream change it was based on has vanished entirely
from Andy's tree. So simply reverting this from ours.

Fixes: 3364daccf32f93e1e13ff5c6ad07ea85a414204a

BUG=chromium:795946
TEST=Boot kernel with Cheza

Change-Id: Ibfbf67a8c4713de9ecb4ee8ddf6d8d4531f020ae
Signed-off-by: Evan Green <evgren@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1058600
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>