FROMLIST: Fix zero-copy path in the 9p virtio transport
The zero-copy optimization when reading or writing large chunks of data
is quite useful.  However, the 9p messages created through the zero-copy
write path have an incorrect message size: it should be the size of the
header + size of the data being written but instead it's just the size
of the header.

This only works if the server ignores the size field of the message and
otherwise breaks the framing of the protocol. Fix this by re-writing the
message size field with the correct value.

Tested by running `dd if=/dev/zero of=out bs=4k count=1` inside a
virtio-9p mount.

BUG=chromium:703939
TEST=`dd if=/dev/zero of=out bs=4k count=1` inside a virtio-9p mount

Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Greg Kurz <groug@kaod.org>
Tested-by: Greg Kurz <groug@kaod.org>

Pulled from https://lkml.org/lkml/2018/7/16/1120

Change-Id: I6e9c22159f1c2b25d0f762a26cf8fa28292e625b
Reviewed-on: https://chromium-review.googlesource.com/1065170
Commit-Ready: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>