UPSTREAM: HID: rmi: use HID_QUIRK_NO_INPUT_SYNC
When we receive a RMI4 report, we should not unconditionally send an
input_sync event. Instead, we should let the rmi4 transport layer do it
for us.

This fixes a situation where we might receive X in a report and the rest
in a subsequent one. And this messes up user space.

Link: https://bugs.freedesktop.org/show_bug.cgi?id=100436

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Oscar Morante <spacepluk@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit c94ba060112ad24fa29b2bdafc0c32173e1f1959)

BUG=b:181300344
TEST=Compile, upload the kernel to Lulu chromebook. Verify if the issue
with jittering touchpad does not occur.

Signed-off-by: Marek Maslanka <mm@semihalf.com>
Change-Id: I5a13213eae8ba23711179a02e4bb61d573607692
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2823790
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>
Reviewed-by: Alex Levin <levinale@google.com>